fix: notes at start of line were not replaced
The regex was expecting a space to precede the lilypond note. However, a note at the very start of a line will not have a leading space. We still know that it is a note because there cannot be anything before the start of a line.
The regex was extended to match both a start of line, and the leading spaces.