feat: find and set last chord
There is a limitation when it comes to multivoices. When inserting right after a multi-voice passage, the last chord will be used (from the last voice in the LilyPond file). However, this is most likely not leading to the desired results.
Regex alone cannot provide a solution to this. We could use more elaborate vim functions to move the cursor around for example. The most ideal scenario would be a treesitter parser for LilyPond, which will make it very clear which voice is which.