feat!: Healthcheck for lilypond-midi-input 0.10
The backend introduces the ability to specify the language in which to
output the LilyPond notes. In order to bring this functionality to this
plugin, it is required to update lilypond-midi-input.

Also introduces option check for the new language option.

See: https://github.com/niveK77pur/lilypond-midi-input/pull/4

BREAKING CHANGE: Requires updating lilypond-midi-input to 0.10